move StoreDependencyMixin to dependencies
