+++ src/dependencies/__tests__/StoreDependencyMixin-test.js
